refactor(scanner): decompose diagram-gen.py into diagram_data + diagram_svg leaf modules
Priority-2 backlog item #2 (docs/plans/refactoring-backlog.md). Pure,
behavior-preserving move — no logic change — to bring diagram-gen.py back under
the 800-line coding-style soft max.

- Extract the scan-data layer to `diagram_data.py`: `CATEGORIES` +
  `_parse_ocsf_json` / `load_prowler_files` / `load_scan_history` /
  `aggregate_scan_data`.
- Extract the SVG builders to `diagram_svg.py`: `_svg_escape` /
  `generate_architecture_svg` / `generate_overview_svg`.
- Re-export both back into `diagram-gen.py` (mirrors the existing `diagram_drawio`
  re-export pattern) so the importlib test harness resolves `MOD.<name>` as before.
- `generate_security_domains_diagram` and the two canonical imports stay in
  diagram-gen.py so the `test_ci_diagram_gen_canonical_sync` frameworks-derive
  and ARCH_DOMAINS guards keep firing.
- Drop now-unused `glob` / `defaultdict` imports; `hashlib` stays (network page).

diagram-gen.py: 838 -> 545 lines.

Verification (offline):
- pytest scanner/tests --cov=scanner/lib --cov-fail-under=99 -> 99.12%, 1477 passed
  (5 pre-existing test_scan_report_baseline failures are a stale local
  scan-report.json, unrelated — fail identically with this change stashed).
- 129 diagram + canonical-sync tests pass.
- All six generated .drawio/.svg files are byte-identical (SHA-256) vs origin/main.
